requirements.txt
# Core packages
pyspark==3.3.1
kafka-python==2.0.2
confluent-kafka==2.0.2
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
pydantic==1.10.7
influxdb

# Data generation
Faker==18.3.1
pendulum==2.1.2

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Utils
python-dotenv==1.0.0
PyYAML==6.0
requests==2.29.0
loguru==0.7.0

# Development
black==23.3.0
flake8==6.0.0
mypy==1.2.0